Resolve issues reported in review, create `RandomTeleportRadiusRepresenterImpl#of` methods.
